rename pkg/weights to pkg/weightslegacy
Free up the pkg/weights import path for the new managed weights
implementation without removing the existing code yet.
