+++ pkg/dockerfile/generator.go
@@ -3,7 +3,7 @@ package dockerfile
 import (
 	"context"
 
-	"github.com/replicate/cog/pkg/weights"
+	"github.com/replicate/cog/pkg/weightslegacy"
 )
 
 type Generator interface {
@@ -17,7 +17,7 @@ type Generator interface {
 	SetUseCudaBaseImage(string)
 	IsUsingCogBaseImage() bool
 	BaseImage(ctx context.Context) (string, error)
-	GenerateWeightsManifest(ctx context.Context) (*weights.Manifest, error)
+	GenerateWeightsManifest(ctx context.Context) (*weightslegacy.Manifest, error)
 	GenerateDockerfileWithoutSeparateWeights(ctx context.Context) (string, error)
 	GenerateModelBase(ctx context.Context) (string, error)
 	Name() string
