+++ pkg/weightslegacy/manifest.go
@@ -1,4 +1,4 @@
-package weights
+package weightslegacy
 
 import (
 	"encoding/binary"
