+++ pkg/weightslegacy/weights_test.go
@@ -1,4 +1,4 @@
-package weights
+package weightslegacy
 
 import (
 	"os"
